+++ changelog.md
@@ -33,6 +33,7 @@
   - The more goombas attached to the player, the slower you move
 - Added tooltips to warp pipes
 - Added advancements
+- Added a config option to give fire & ice flower power ups a duration. (Disabled by default)
 - Added an optional `durationTicks` field to the power up command for fire & ice power ups
 - Added new spawn animations to power ups
 - Added flame particles to the Fire Flower